move AnnotatedPropertyInjectToConstructorInjectionRector to general rectors
