--- packages/Nette/tests/Rector/Application/AnnotatedPropertyInjectToConstructorInjectionRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Nette\Rector\Application\AnnotatedPropertyInjectToConstructorInjectionRector: ~
